+++ packages/renderer/src/lib/pod/PodCreateFromContainers.svelte
@@ -12,10 +12,10 @@ import EngineFormPage from '/@/lib/ui/EngineFormPage.svelte';
 import WarningMessage from '/@/lib/ui/WarningMessage.svelte';
 import { handleNavigation } from '/@/navigation';
 import { providerInfos } from '/@/stores/providers';
+import type { PodCreatePortOptions } from '/@api/libpod/libpod';
 import { NavigationPage } from '/@api/navigation-page';
 import type { ProviderContainerConnectionInfo, ProviderInfo } from '/@api/provider-info';
 
-import type { PodCreatePortOptions } from '../../../../main/src/plugin/dockerode/libpod-dockerode';
 import { type PodCreation, podCreationHolder } from '../../stores/creation-from-containers-store';
 
 let podCreation: PodCreation;
